migrated header and footer to emotion
