+++ .babelrc
@@ -12,6 +12,12 @@
   },
   "plugins": [
     [
+      "emotion",
+      {
+        "inline": true
+      }
+    ],
+    [
       "lodash",
       {
         "id": ["lodash", "semantic-ui-react"]
