+++ components/common/layout.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import Header from './header';
+import Footer from './footer';
+
+export default ({ children }) => (
+  <div>
+    <Header />
+    {children}
+    <Footer />
+  </div>
+);
