+++ package.json
@@ -53,7 +53,11 @@
     ]
   },
   "dependencies": {
+    "babel-plugin-emotion": "^8.0.10",
     "date-fns": "1.29.0",
+    "emotion": "^8.0.10",
+    "emotion-server": "^8.0.10",
+    "grid-emotion": "^2.1.0",
     "isomorphic-unfetch": "2.0.0",
     "lodash.take": "^4.1.1",
     "marked": "^0.3.6",
@@ -63,9 +67,11 @@
     "prop-types": "^15.5.10",
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
+    "react-emotion": "^8.0.10",
     "react-headroom": "^2.1.6",
     "react-icons": "^2.2.5",
-    "semantic-ui-react": "^0.76.0"
+    "semantic-ui-react": "^0.76.0",
+    "styled-system": "^1.0.8"
   },
   "devDependencies": {
     "babel-eslint": "^8.0.1",
